requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Slack SDK
slack-sdk>=3.26.0

# Google Gemini API
google-generativeai>=0.4.0

# PDF Processing
pdfplumber>=0.10.0

# Configuration
pyyaml>=6.0

# HTTP client (for PDF downloads)
requests>=2.31.0
